+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# dev
+# 1.14.0
 - color: add vegetation_dim parameter
 - las_add_buffer: fix test checking input and output dimensions consistency
 - [new feature] replace_area_in_pointcloud: replace part of a pointcloud by the points of another pointcloud